requirements.txt
instaloader==4.10.2
brevo-python==1.0.0
python-dotenv==1.0.0
requests==2.31.0
